feat: Add GridSearchCV to optimize hyperparameters
Use GridSearchCV to try different value combinations of some of the hyperparameters in RandomForestClassifier during training, this way the model can adapt better to new training data. Because this increased the training time by a significant factor, I also set `n_jobs=-1` on the RandomForestClassifier itself so cores usage is maximized. I added a warning to the readme to set expectations for the user based on my experience.